[FIX] tests fatturapa in
