+++ l10n_it_fatturapa_in/tests/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 
 from . import test_import_fatturapa_xml
+from . import test_import_fatturapa_xml1
+from . import test_import_fatturapa_xml2
