+++ l10n_it_fatturapa_in/tests/test_import_fatturapa_xml.py
@@ -155,29 +155,6 @@ class TestFatturaPAXMLValidation(SingleTransactionCase):
         self.assertEqual(invoice.fiscal_document_type_id.code, 'TD01')
         self.assertTrue(invoice.art73)
 
-    # def test_02_xml_import(self):
-    #     res = self.run_wizard('test2', 'IT03638121008_X11111.xml')
-    #     invoice_id = res.get('domain')[0][2][0]
-    #     invoice = self.invoice_model.browse(invoice_id)
-    #     self.assertEqual(invoice.supplier_invoice_number, '00001')
-    #     self.assertEqual(invoice.amount_untaxed, 3)
-    #     self.assertEqual(invoice.amount_tax, 0.66)
-    #     self.assertEqual(
-    #         invoice.fatturapa_summary_ids[0].amount_untaxed, 3)
-    #     self.assertEqual(
-    #         invoice.fatturapa_summary_ids[0].amount_tax, 0.66)
-    #     self.assertEqual(invoice.partner_id.name, "Societa' alpha S.r.l.")
-
-    # def test_03_xml_import(self):
-    #     res = self.run_wizard('test3', 'IT05979361218_002.xml.p7m')
-    #     invoice_id = res.get('domain')[0][2][0]
-    #     invoice = self.invoice_model.browse(invoice_id)
-    #     self.assertEqual(invoice.partner_id.register_code, 'TO1258B')
-    #     self.assertEqual(
-    #         invoice.partner_id.register_fiscalpos.code, 'RF02')
-    #     self.assertEqual(invoice.supplier_invoice_number, 'FT/2015/0007')
-    #     self.assertEqual(invoice.amount_total, 54.00)
-
     def test_04_xml_import(self):
         res = self.run_wizard('test4', 'IT02780790107_11005.xml')
         invoice_id = res.get('domain')[0][2][0]
@@ -208,121 +185,6 @@ class TestFatturaPAXMLValidation(SingleTransactionCase):
             invoice.inconsistencies,
             u'DatiAnagrafici.Anagrafica.Denominazione contains "Societa\' '
             'Alpha SRL". Your System contains "SOCIETA\' ALPHA SRL"\n\n')
-
-    def test_05_xml_import(self):
-        self.env.user.company_id.dati_bollo_product_id = (
-            self.service.id)
-        res = self.run_wizard('test5', 'IT05979361218_003.xml')
-        invoice_id = res.get('domain')[0][2][0]
-        invoice = self.invoice_model.browse(invoice_id)
-        self.assertEqual(invoice.reference, 'FT/2015/0008')
-        self.assertEqual(invoice.sender, 'TZ')
-        self.assertEqual(invoice.intermediary.name, 'ROSSI MARIO')
-        self.assertEqual(invoice.intermediary.firstname, 'MARIO')
-        self.assertEqual(invoice.intermediary.lastname, 'ROSSI')
-        bollo_found = False
-        for line in invoice.invoice_line:
-            if line.product_id.id == self.service.id:
-                self.assertEqual(line.price_unit, 6)
-                bollo_found = True
-        self.assertTrue(bollo_found)
-        self.assertEqual(
-            invoice.e_invoice_line_ids[0].discount_rise_price_ids[0].name,
-            'SC')
-        self.assertEqual(
-            invoice.e_invoice_line_ids[0].discount_rise_price_ids[0].
-            percentage, 10
-        )
-        self.assertEqual(invoice.amount_untaxed, 15)
-        self.assertEqual(invoice.amount_tax, 0)
-        self.assertEqual(invoice.amount_total, 15)
-
-    def test_06_import_except(self):
-        # File not exist Exception
-        self.assertRaises(
-            Exception, self.run_wizard, 'test6_Exception', '')
-        # fake Signed file is passed , generate orm_exception
-        self.assertRaises(
-            UserError, self.run_wizard, 'test6_orm_exception',
-            'IT05979361218_fake.xml.p7m'
-        )
-
-    def test_07_xml_import(self):
-        # 2 lines with quantity != 1 and discounts
-        res = self.run_wizard('test7', 'IT05979361218_004.xml')
-        invoice_id = res.get('domain')[0][2][0]
-        invoice = self.invoice_model.browse(invoice_id)
-        self.assertEqual(invoice.reference, 'FT/2015/0009')
-        self.assertAlmostEqual(invoice.amount_untaxed, 1173.60)
-        self.assertEqual(invoice.amount_tax, 258.19)
-        self.assertEqual(invoice.amount_total, 1431.79)
-        self.assertEqual(invoice.invoice_line[0].admin_ref, 'D122353')
-
-    def test_08_xml_import(self):
-        # using ImportoTotaleDocumento
-        res = self.run_wizard('test8', 'IT05979361218_005.xml')
-        invoice_id = res.get('domain')[0][2][0]
-        invoice = self.invoice_model.browse(invoice_id)
-        self.assertEqual(invoice.reference, 'FT/2015/0010')
-        self.assertAlmostEqual(invoice.amount_total, 1288.61)
-        self.assertFalse(invoice.inconsistencies)
-
-    def test_09_xml_import(self):
-        # using DatiGeneraliDocumento.ScontoMaggiorazione without
-        # ImportoTotaleDocumento
-        # add test file name case sensitive
-        res = self.run_wizard('test9', 'IT05979361218_006.XML')
-        invoice_id = res.get('domain')[0][2][0]
-        invoice = self.invoice_model.browse(invoice_id)
-        self.assertEqual(invoice.reference, 'FT/2015/0011')
-        self.assertAlmostEqual(invoice.amount_total, 1288.61)
-        self.assertEqual(
-            invoice.inconsistencies,
-            'Computed amount untaxed 1030.42 is different from'
-            ' DatiRiepilogo 1173.6')
-
-    def test_10_xml_import(self):
-        # Fix Date format
-        res = self.run_wizard('test6', 'IT05979361218_007.xml')
-        invoice_id = res.get('domain')[0][2][0]
-        invoice = self.invoice_model.browse(invoice_id)
-        self.assertEqual(invoice.reference, 'FT/2015/0009')
-        self.assertEqual(
-            invoice.date_invoice, '2015-03-16')
-        self.assertEqual(
-            invoice.fatturapa_payments[0].payment_methods[0].payment_due_date,
-            '2015-06-03'
-        )
-        self.assertEqual(
-            invoice.fatturapa_payments[0].payment_methods[0].
-            fatturapa_pm_id.code,
-            'MP18'
-        )
-
-    def test_11_xml_import(self):
-        # DatiOrdineAcquisto with RiferimentoNumeroLinea referring to
-        # not existing invoice line
-        res = self.run_wizard('test11', 'IT02780790107_11006.xml')
-        invoice_id = res.get('domain')[0][2][0]
-        invoice = self.invoice_model.browse(invoice_id)
-        self.assertEqual(
-            len(invoice.invoice_line[0].related_documents), 0)
-        self.assertEqual(
-            invoice.invoice_line[0].sequence, 1)
-        self.assertEqual(
-            invoice.related_documents[0].type, "order")
-        self.assertEqual(
-            invoice.related_documents[0].lineRef, 60)
-
-    def test_12_xml_import(self):
-        res = self.run_wizard('test12', 'IT05979361218_008.xml')
-        invoice_id = res.get('domain')[0][2][0]
-        invoice = self.invoice_model.browse(invoice_id)
-        self.assertEqual(invoice.reference, 'FT/2015/0012')
-        self.assertEqual(invoice.sender, 'TZ')
-        self.assertEqual(invoice.intermediary.name, 'ROSSI MARIO')
-        self.assertEqual(invoice.intermediary.firstname, 'MARIO')
-        self.assertEqual(invoice.intermediary.lastname, 'ROSSI')
 
     def test_13_xml_import(self):
         # inconsistencies must not be duplicated
@@ -361,15 +223,6 @@ class TestFatturaPAXMLValidation(SingleTransactionCase):
             ' but it does not exist in your system\n'
             'XML contains tax with percentage "15.55"'
             ' but it does not exist in your system')
-
-    def test_15_xml_import(self):
-        self.wt = self.create_wt()
-        res = self.run_wizard('test15', 'IT05979361218_009.xml')
-        invoice_id = res.get('domain')[0][2][0]
-        invoice = self.invoice_model.browse(invoice_id)
-        self.assertAlmostEquals(invoice.withholding_tax_amount, 1)
-        self.assertAlmostEquals(invoice.amount_total, 6.1)
-        self.assertAlmostEquals(invoice.amount_net_pay, 5.1)
 
     def test_16_xml_import(self):
         # file B2B downloaded from
