+++ l10n_it_fatturapa_in/tests/test_import_fatturapa_xml2.py
@@ -1,0 +1,139 @@
+# -*- coding: utf-8 -*-
+
+import base64
+import tempfile
+from openerp.tests.common import SingleTransactionCase
+from openerp.modules import get_module_resource
+from openerp.exceptions import Warning as UserError
+
+
+class TestFatturaPAXMLValidation(SingleTransactionCase):
+
+    def getFile(self, filename):
+        path = get_module_resource(
+            'l10n_it_fatturapa_in', 'tests', 'data', filename)
+        with open(path) as test_data:
+            with tempfile.TemporaryFile() as out:
+                base64.encode(test_data, out)
+                out.seek(0)
+                return path, out.read()
+
+    def create_wt(self):
+        return self.env['withholding.tax'].create({
+            'name': '1040',
+            'account_receivable_id': self.payable_account_id,
+            'account_payable_id': self.payable_account_id,
+            'payment_term': self.env.ref('account.account_payment_term').id,
+            'rate_ids': [(0, 0, {'tax': 20.0})],
+            'causale_pagamento_id':
+                self.env.ref('l10n_it_causali_pagamento.a').id,
+        })
+
+    def setUp(self):
+        super(TestFatturaPAXMLValidation, self).setUp()
+        self.wizard_model = self.env['wizard.import.fatturapa']
+        self.data_model = self.env['ir.model.data']
+        self.attach_model = self.env['fatturapa.attachment.in']
+        self.invoice_model = self.env['account.invoice']
+        self.payable_account_id = self.env['account.account'].search([
+            ('user_type', '=', self.env.ref(
+                'account.data_account_type_payable').id)
+        ], limit=1).id
+        self.headphones = self.env.ref(
+            'product.product_product_7_product_template')
+        self.imac = self.env.ref(
+            'product.product_product_8_product_template')
+        self.service = self.env.ref('product.product_product_consultant')
+        self.state = self.env['res.country.state']
+        self.state.create({
+            'code': 'SS',
+            'name': 'Sassari',
+            'country_id': self.env['res.country'].search(
+                [('code', '=', 'IT')]).id
+        })
+
+    def run_wizard(self, name, file_name):
+        attach_id = self.attach_model.create(
+            {
+                'name': name,
+                'datas': self.getFile(file_name)[1],
+                'datas_fname': file_name
+            }).id
+        wizard = self.wizard_model.with_context(
+            active_ids=[attach_id]).create({})
+        return wizard.importFatturaPA()
+
+    def test_08_xml_import(self):
+        # using ImportoTotaleDocumento
+        res = self.run_wizard('test8', 'IT05979361218_005.xml')
+        invoice_id = res.get('domain')[0][2][0]
+        invoice = self.invoice_model.browse(invoice_id)
+        self.assertEqual(invoice.reference, 'FT/2015/0010')
+        self.assertAlmostEqual(invoice.amount_total, 1288.61)
+        self.assertFalse(invoice.inconsistencies)
+
+    def test_09_xml_import(self):
+        # using DatiGeneraliDocumento.ScontoMaggiorazione without
+        # ImportoTotaleDocumento
+        # add test file name case sensitive
+        res = self.run_wizard('test9', 'IT05979361218_006.XML')
+        invoice_id = res.get('domain')[0][2][0]
+        invoice = self.invoice_model.browse(invoice_id)
+        self.assertEqual(invoice.reference, 'FT/2015/0011')
+        self.assertAlmostEqual(invoice.amount_total, 1288.61)
+        self.assertEqual(
+            invoice.inconsistencies,
+            'Computed amount untaxed 1030.42 is different from'
+            ' DatiRiepilogo 1173.6')
+
+    def test_10_xml_import(self):
+        # Fix Date format
+        res = self.run_wizard('test6', 'IT05979361218_007.xml')
+        invoice_id = res.get('domain')[0][2][0]
+        invoice = self.invoice_model.browse(invoice_id)
+        self.assertEqual(invoice.reference, 'FT/2015/0009')
+        self.assertEqual(
+            invoice.date_invoice, '2015-03-16')
+        self.assertEqual(
+            invoice.fatturapa_payments[0].payment_methods[0].payment_due_date,
+            '2015-06-03'
+        )
+        self.assertEqual(
+            invoice.fatturapa_payments[0].payment_methods[0].
+            fatturapa_pm_id.code,
+            'MP18'
+        )
+
+    def test_11_xml_import(self):
+        # DatiOrdineAcquisto with RiferimentoNumeroLinea referring to
+        # not existing invoice line
+        res = self.run_wizard('test11', 'IT02780790107_11006.xml')
+        invoice_id = res.get('domain')[0][2][0]
+        invoice = self.invoice_model.browse(invoice_id)
+        self.assertEqual(
+            len(invoice.invoice_line[0].related_documents), 0)
+        self.assertEqual(
+            invoice.invoice_line[0].sequence, 1)
+        self.assertEqual(
+            invoice.related_documents[0].type, "order")
+        self.assertEqual(
+            invoice.related_documents[0].lineRef, 60)
+
+    def test_12_xml_import(self):
+        res = self.run_wizard('test12', 'IT05979361218_008.xml')
+        invoice_id = res.get('domain')[0][2][0]
+        invoice = self.invoice_model.browse(invoice_id)
+        self.assertEqual(invoice.reference, 'FT/2015/0012')
+        self.assertEqual(invoice.sender, 'TZ')
+        self.assertEqual(invoice.intermediary.name, 'ROSSI MARIO')
+        self.assertEqual(invoice.intermediary.firstname, 'MARIO')
+        self.assertEqual(invoice.intermediary.lastname, 'ROSSI')
+
+    def test_15_xml_import(self):
+        self.wt = self.create_wt()
+        res = self.run_wizard('test15', 'IT05979361218_009.xml')
+        invoice_id = res.get('domain')[0][2][0]
+        invoice = self.invoice_model.browse(invoice_id)
+        self.assertAlmostEquals(invoice.withholding_tax_amount, 1)
+        self.assertAlmostEquals(invoice.amount_total, 6.1)
+        self.assertAlmostEquals(invoice.amount_net_pay, 5.1)
