+++ l10n_it_fatturapa_in/wizard/link_to_existing_invoice.xml
@@ -12,6 +12,7 @@
                             <field name="invoice_id"
                                    domain="[('type', 'in', ('in_invoice', 'in_refund')), ('state', 'in', ('draft', 'proforma', 'proforma2')), ('fatturapa_attachment_in_id', '=', False)]"
                                    context="{'form_view_ref': 'account.invoice_supplier_form', 'tree_view_ref': 'account.invoice_supplier_tree'}"
+                                   options="{'no_create': True}"
                             />
                         </group>
                         <footer>
